Split representation of per-variable and shared aspects of variable declarations (#23)
Introduce a `VariableDeclarationGroup` for `VariableDeclaration`s that share the same type and modifiers.

This is a similar pattern to the existing `FieldDeclarationGroup` and `FieldDeclaration` nodes.

With this, we properly separate the source code positions of the per-variable and the shared syntax.
Some use cases depend on this (specifically, the `OneDeclarationPerLine` rule).

Retype and rename some various class members that are now a `VariableDeclarationGroup`. Add a
convenience method to construct a unitary group, which is a common case.

Update various unit tests.