Reorder class body declarations (so that properties are visited before methods).
This is a semi-unprincipled change that is sufficient to fix PMD tests.
We may want to revisit later whether class body declarations -- or even
all children in general --  should be traversed in the same order as they
appear in the source code. (This is somewhat tricky to do within the constraints
imposed by the serialization approach.)

Add unit test.

Change-Id: Ie1e4546fca8409c95052c16aeeb35feb149dda65