Create an embryo symbol resolver.
This changes `SummitTool` to resolve the symbols after it parses the
Apex code. Notes:

1. `resolve` does not yet return anything

2. `resolve` currently only handles (collects) class definitions and
their methods.

3. Turn `SummitResolver.PRINT_DEBUG` to `true` to see a printout of
what gets collected.

4. The printout described in #3 uses standard Kotlin `print`, rather
than a `FluentLogger`. The output of the `FluentLogger` are too messy
for that, and it does not really seem like the right thing to use.

5. Similiarly to #4, it does not seem right that `SummitTool.kt`
should use a `FluentLogger` either, and a TODO was added to it.

6. Another TODO was added to `SummitTool.kt` to support flags